test: define targets in conftest.py, refactor devtools usages
* configure dict mapping exe names to paths in conftest.py instead of importing from devtools (to be removed)
* remove usages of Case (to be removed from devtools), refactor test_gwf_maw[1-4] to accommodate
