+++ autotest/test_gwf_maw01.py
@@ -1,47 +1,57 @@
-# Test for checking maw observation input.  The following observation types:
-# 'maw' and 'conductance,' require that ID2 be provided when
-# ID is an integer corresponding to a well number and not BOUNDNAME.
-# See table in MAW Package section of mf6io.pdf for an explanation of ID,
-# ID2, and Observation Type.
-
-
 import os
+from types import SimpleNamespace as Case
 
 import flopy
+import numpy as np
+import pytest
 
-newtonoptions = [None, "NEWTON", "NEWTON UNDER_RELAXATION"]
-ex = "maw_obs"
+from framework import TestFramework
+from simulation import TestSimulation
+
+budtol = 1e-2
+bud_lst = ["GWF_IN", "GWF_OUT", "RATE_IN", "RATE_OUT"]
+
+well1 = Case(
+    observations={"maw_obs.csv": [("mh1", "head", 1)]},
+    packagedata=[[0, 0.1, 50.0, 100.0, "THIEM", 1]],
+    connectiondata=[[0, 0, (0, 0, 1), 100.0, 50.0, 1.0, 0.1]],
+    perioddata=[[0, "rate", 0.0]],
+)
+
+ex = ["maw01", "maw01nwt", "maw01nwtur"]
+krylov = ["CG", "BICGSTAB", "BICGSTAB"]
+newton = [None, "NEWTON", "NEWTON UNDER_RELAXATION"]
+nlay = 1
+nrow = 1
+ncol = 3
+nper = 3
+delr = 300
+delc = 300
+perlen = 3 * [1]
+nstp = 3 * [1]
+tsmult = 3 * [1]
+well = well1
+strt = 100
+hk = 1
+nouter = 100
+ninner = 300
+hclose = 1e-9
+rclose = 1e-3
+relaxation_factor = 1
+compare = False
 
 
-def build_model(dir, exe):
-    nlay, nrow, ncol = 1, 1, 3
-    nper = 3
-    perlen = [1.0, 1.0, 1.0]
-    nstp = [1, 1, 1]
-    tsmult = [1.0, 1.0, 1.0]
-    lenx = 300.0
-    delr = delc = lenx / float(nrow)
-    strt = 100.0
-    hnoflo = 1e30
-    hdry = -1e30
-    hk = 1.0
-
-    nouter, ninner = 100, 300
-    hclose, rclose, relax = 1e-9, 1e-3, 1.0
-    krylov = ["CG", "BICGSTAB", "BICGSTAB"]
-
-    tdis_rc = []
-    for i in range(nper):
-        tdis_rc.append((perlen[i], nstp[i], tsmult[i]))
-
-    name = ex
-
-    # build MODFLOW 6 files
-    ws = dir
+def build_model(idx, ws, mf6):
+    name = ex[idx]
     sim = flopy.mf6.MFSimulation(
-        sim_name=name, version="mf6", exe_name=exe, sim_ws=ws
+        sim_name=name,
+        version="mf6",
+        exe_name=mf6,
+        sim_ws=ws,
     )
+
     # create tdis package
+    tdis_rc = [(perlen[i], nstp[i], tsmult[i]) for i in range(nper)]
     tdis = flopy.mf6.ModflowTdis(
         sim, time_units="DAYS", nper=nper, perioddata=tdis_rc
     )
@@ -53,7 +63,7 @@ def build_model(dir, exe):
         modelname=name,
         model_nam_file=f"{name}.nam",
     )
-    gwf.name_file.newtonoptions = newtonoptions[0]
+    gwf.name_file.newtonoptions = newton[idx]
 
     # create iterative model solution and register the gwf model with it
     ims = flopy.mf6.ModflowIms(
@@ -65,10 +75,10 @@ def build_model(dir, exe):
         inner_maximum=ninner,
         inner_dvclose=hclose,
         rcloserecord=rclose,
-        linear_acceleration=krylov[0],
+        linear_acceleration=krylov[idx],
         scaling_method="NONE",
         reordering_method="NONE",
-        relaxation_factor=relax,
+        relaxation_factor=relaxation_factor,
     )
     sim.register_ims_package(ims, [gwf.name])
 
@@ -132,13 +142,6 @@ def build_model(dir, exe):
     #                              periodrecarray=wd6,
     #                              save_flows=False)
     # MAW
-    opth = f"{name}.maw.obs"
-    wellbottom = 50.0
-    wellrecarray = [[0, 0.1, wellbottom, 100.0, "THIEM", 1]]
-    wellconnectionsrecarray = [[0, 0, (0, 0, 1), 100.0, wellbottom, 1.0, 0.1]]
-    wellperiodrecarray = [[0, "rate", 0.0]]
-    mawo_dict = {}
-    mawo_dict["maw_obs.csv"] = [("mh1", "head", 1), ("mawgw", "maw", 1)]
     maw = flopy.mf6.ModflowGwfmaw(
         gwf,
         filename=f"{name}.maw",
@@ -146,10 +149,10 @@ def build_model(dir, exe):
         print_head=True,
         print_flows=True,
         save_flows=True,
-        observations=mawo_dict,
-        packagedata=wellrecarray,
-        connectiondata=wellconnectionsrecarray,
-        perioddata=wellperiodrecarray,
+        observations=well.observations,
+        packagedata=well.packagedata,
+        connectiondata=well.connectiondata,
+        perioddata=well.perioddata,
     )
 
     # output control
@@ -163,51 +166,48 @@ def build_model(dir, exe):
         filename=f"{name}.oc",
     )
 
-    return sim
+    return sim, None
 
 
-def test_mf6model(function_tmpdir, targets):
-    mf6 = targets["mf6"]
-    sim = build_model(str(function_tmpdir), mf6)
+def eval_results(sim):
+    print("evaluating MAW heads...")
+
+    # MODFLOW 6 maw results
+    fpth = os.path.join(sim.simpath, "maw_obs.csv")
+    tc = np.genfromtxt(fpth, names=True, delimiter=",")
+
+    # create known results array
+    tc0 = np.array([100.0, 25.0, 100.0])
+
+    # calculate maximum absolute error
+    diff = tc["MH1"] - tc0
+    diffmax = np.abs(diff).max()
+    dtol = 1e-9
+    msg = f"maximum absolute maw head difference ({diffmax}) "
+
+    if diffmax > dtol:
+        sim.success = False
+        msg += f"exceeds {dtol}"
+        assert diffmax < dtol, msg
+    else:
+        sim.success = True
+        print("    " + msg)
+
+
+@pytest.mark.parametrize("idx, name", list(enumerate(ex)))
+def test_mf6model(idx, name, function_tmpdir, targets):
+    ws = str(function_tmpdir)
+    sim, _ = build_model(idx, ws, targets.mf6)
     sim.write_simulation()
     sim.run_simulation()
-
-    # ensure that the error msg is contained in the mfsim.lst file
-    f = open(str(function_tmpdir / "mfsim.lst"), "r")
-    lines = f.readlines()
-    error_count = 0
-    expected_msg = False
-    for line in lines:
-        if "ID2 (icon) is missing" in line:
-            expected_msg = True
-            error_count += 1
-
-    assert error_count == 1, (
-        "error count = " + str(error_count) + ", but should equal 1"
+    test = TestFramework()
+    test.run(
+        TestSimulation(
+            name=name,
+            exe_dict=targets,
+            exfunc=eval_results,
+            idxsim=idx,
+            make_comparison=False,
+        ),
+        ws,
     )
-
-    # fix the error and attempt to rerun model
-    orig_fl = str(function_tmpdir / (ex + ".maw.obs"))
-    new_fl = str(function_tmpdir / (ex + ".maw.obs.new"))
-    sr = open(orig_fl, "r")
-    sw = open(new_fl, "w")
-
-    lines = sr.readlines()
-    error_free_line = "  mawgw  maw  1  1\n"
-    for line in lines:
-        if " maw " in line:
-            sw.write(error_free_line)
-        else:
-            sw.write(line)
-
-    sr.close()
-    sw.close()
-
-    # delete original and replace with corrected lab obs input
-    os.remove(orig_fl)
-    os.rename(new_fl, orig_fl)
-
-    # rerun the model, should be no errors
-    success, buff = sim.run_simulation()
-
-    assert success, "model rerun failed"
