+++ autotest/test_gwf_maw06.py
@@ -7,6 +7,7 @@ import os
 import flopy
 import numpy as np
 import pytest
+
 from framework import TestFramework
 from simulation import TestSimulation
