+++ autotest/test_gwf_maw08.py
@@ -6,6 +6,7 @@ import os
 import flopy
 import numpy as np
 import pytest
+
 from framework import TestFramework
 from simulation import TestSimulation
